+++ lib/wrapture/scope.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Wrapture
+  # Describes a scope of one or more class specifications.
+  class Scope
+  end
+end